refactor: split cleanupStaleWindowsOldExe into windows-tagged file
The windows-only .old cleanup lived in root.go behind a runtime.GOOS
guard, so the 7 windows-only lines counted as patch-misses on the
ubuntu coverage run (which can never execute them). Move the real
impl into cleanup_windows.go (//go:build windows) and a no-op stub
into cleanup_unix.go (//go:build !windows), mirroring how the project
already isolates internal/desktop/app.go. root.go's runtime import is
dropped (no longer needed). The windows-tagged test in
root_windows_test.go now covers the real impl on the Windows runner.
