+++ cmd/javinizer/root.go
@@ -5,7 +5,6 @@ import (
 	"net/url"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strconv"
 
 	"github.com/javinizer/javinizer-go/cmd/javinizer/commands/actress"
@@ -65,24 +64,9 @@ func initDesktopDefault() {
 }
 
 // cleanupStaleWindowsOldExe removes a leftover <exe>.old from a prior
-// interrupted desktop self-upgrade. Windows can't overwrite a running exe, so
-// the upgrade renames it to .old and the detached helper deletes it after the
-// swap; if the helper was killed before cleanup, the .old survives and is
-// cleared here on the next launch. No-op on non-Windows.
-func cleanupStaleWindowsOldExe() {
-	if runtime.GOOS != "windows" {
-		return
-	}
-	exe, err := os.Executable()
-	if err != nil {
-		return
-	}
-	if resolved, err := filepath.EvalSymlinks(exe); err == nil {
-		_ = os.Remove(resolved + ".old")
-	} else {
-		_ = os.Remove(exe + ".old")
-	}
-}
+// interrupted desktop self-upgrade (Windows can't overwrite a running exe).
+// The Windows implementation lives in cleanup_windows.go; on other platforms
+// it is a no-op (cleanup_unix.go).
 
 func init() {
 	// Customize version template
